Добавляет модульность JS
